Rename `multirate_taps()` to `design_multirate_fir()`
Closes Rename `sdr.multirate_taps()` to `sdr.design_multirate_fir()` #151
