+++ src/sdr/_filter/__init__.py
@@ -2,6 +2,7 @@
 A subpackage for various LTI filters.
 """
 from ._design_fir import *
+from ._design_multirate import *
 from ._fir import *
 from ._iir import *
 from ._polyphase import *
