+++ src/sdr/_filter/_polyphase.py
@@ -3,210 +3,20 @@ A module for interpolating finite impulse response (FIR) filters.
 """
 from __future__ import annotations
 
-import math
-
 import numpy as np
 import numpy.typing as npt
 import scipy.signal
 from typing_extensions import Literal
 
 from .._helper import export
+from ._design_multirate import (
+    design_multirate_fir,
+    design_multirate_fir_linear,
+    design_multirate_fir_linear_matlab,
+    design_multirate_fir_zoh,
+    polyphase_matrix,
+)
 from ._fir import FIR
-
-
-@export
-def multirate_taps(
-    up: int,
-    down: int = 1,
-    half_length: int = 12,
-    A_stop: float = 80,
-) -> npt.NDArray[np.float_]:
-    r"""
-    Computes the multirate FIR filter that achieves rational resampling by $P/Q$.
-
-    Note:
-        This filter can be used with :class:`sdr.Interpolator`, :class:`sdr.Decimator`, or :class:`sdr.Resampler`.
-
-    Arguments:
-        up: The interpolation rate $P$.
-        down: The decimation rate $Q$.
-        half_length: The half-length of the polyphase filters.
-        A_stop: The stopband attenuation $A_{\text{stop}}$ in dB.
-
-    Returns:
-        The multirate FIR filter $h[n]$.
-
-    References:
-        - fred harris, *Multirate Signal Processing for Communication Systems*, Chapter 7: Resampling Filters
-        - https://www.mathworks.com/help/dsp/ref/designmultiratefir.html
-
-    Examples:
-        Design a multirate FIR filter for rational resampling by 11/3.
-
-        .. ipython:: python
-
-            h = sdr.multirate_taps(11, 3)
-
-            @savefig sdr_multirate_fir_1.png
-            plt.figure(figsize=(8, 4)); \
-            sdr.plot.impulse_response(h);
-
-            @savefig sdr_multirate_fir_2.png
-            plt.figure(figsize=(8, 4)); \
-            sdr.plot.magnitude_response(h);
-
-    Group:
-        dsp-multirate-filtering
-    """
-    if not isinstance(up, int):
-        raise TypeError(f"Argument 'up' must be an integer, not {up}.")
-    if not up >= 1:
-        raise ValueError(f"Argument 'up' must be at least 1, not {up}.")
-    P = up
-
-    if not isinstance(down, int):
-        raise TypeError(f"Argument 'down' must be an integer, not {down}.")
-    if not down >= 1:
-        raise ValueError(f"Argument 'down' must be at least 1, not {down}.")
-    Q = down
-
-    B = P if P > 1 else Q  # The number of polyphase branches
-    R = max(P, Q)  # Inverse of the filter's fractional bandwidth
-
-    # Compute the filter order, which is length - 1
-    N = 2 * half_length * B
-
-    # Compute ideal lowpass filter
-    n = np.arange(N + 1)
-    h = P / R * np.sinc((n - N // 2) / R)
-
-    # Compute Kaiser window
-    # beta = scipy.signal.windows.kaiser_beta(A_stop)
-    if A_stop >= 50:
-        beta = 0.1102 * (A_stop - 8.71)  # TODO: MATLAB uses 8.71 and SciPy uses 8.7
-    elif A_stop > 21:
-        beta = 0.5842 * (A_stop - 21) ** 0.4 + 0.07886 * (A_stop - 21)
-    else:
-        beta = 0
-    w = scipy.signal.windows.kaiser(N + 1, beta)
-
-    # Compute windowed filter
-    h = h * w
-
-    if not (Q > P > 1 and (half_length * P) % Q != 0):
-        # The first and last elements are zero. Remove the last zero so that the filter is evenly
-        # partitioned across the polyphase branches. The filter now has length 2 * half_length * B and
-        # each polyphase branch has length 2 * half_length.
-        h = h[:-1]
-
-    # If the above condition is not true, the first and last elements are non-zero. The filter length
-    # is 2 * half_length * B + 1 and each polyphase branch has length 2 * half_length + 1. The final
-    # column in the polyphase matrix will be padded with zeros.
-
-    return h
-
-
-def _multirate_taps_linear(rate: int) -> npt.NDArray[np.float_]:
-    r"""
-    The multirate filter is designed to linearly interpolate between samples. The filter coefficients are a
-    length-$2r$ linear ramp $\frac{1}{r} [0, ..., r-1, r, r-1, ..., 1]$. The first output sample aligns with the
-    first input sample.
-    """
-    h = np.zeros(2 * rate, dtype=float)
-    h[:rate] = np.arange(0, rate) / rate
-    h[rate:] = np.arange(rate, 0, -1) / rate
-    return h
-
-
-def _multirate_taps_linear_matlab(rate: int) -> npt.NDArray[np.float_]:
-    r"""
-    The multirate filter is designed to linearly interpolate between samples. The filter coefficients are a
-    length-$2r$ linear ramp $\frac{1}{r} [1, ..., r-1, r, r-1, ..., 0]$. This is method MATLAB uses. The first
-    output sample is advanced from the first input sample.
-    """
-    h = np.zeros(2 * rate, dtype=float)
-    h[:rate] = np.arange(1, rate + 1) / rate
-    h[rate:] = np.arange(rate - 1, -1, -1) / rate
-    return h
-
-
-def _multirate_taps_zoh(rate: int) -> npt.NDArray[np.float_]:
-    """
-    The multirate filter is designed to be a zero-order hold. The filter coefficients are a length-$r$ array of ones.
-    """
-    h = np.ones(rate, dtype=float)
-    return h
-
-
-@export
-def polyphase_matrix(up: int, down: int, taps: npt.ArrayLike) -> npt.NDArray:
-    """
-    Converts the multirate FIR filter taps $h_i$ into the polyphase matrix $H_{i, j}$ that achieves
-    rational resampling by $P/Q$.
-
-    Arguments:
-        up: The interpolation rate $P$.
-        down: The decimation rate $Q$.
-        taps: The multirate FIR filter taps $h_i$.
-
-    Returns:
-        The polyphase matrix $H_{i, j}$.
-
-    Notes:
-        The multirate FIR filter taps $h_i$ are arranged down the columns of the polyphase matrix
-        $H_{i, j}$ as follows:
-
-        .. code-block:: text
-            :caption: Polyphase Matrix
-
-            +------+------+------+------+
-            | h[0] | h[3] | h[6] | h[9] |
-            +------+------+------+------+
-            | h[1] | h[4] | h[7] | 0    |
-            +------+------+------+------+
-            | h[2] | h[5] | h[8] | 0    |
-            +------+------+------+------+
-
-    References:
-        - fred harris, *Multirate Signal Processing for Communication Systems*, Chapter 7: Resampling Filters
-
-    Examples:
-        Convert the multirate FIR filter (notional taps for demonstration) into a polyphase matrix
-        for rational resampling by 3/4 and 1/6.
-
-        .. ipython:: python
-
-            h = np.arange(0, 20)
-            sdr.polyphase_matrix(3, 4, h)
-            sdr.polyphase_matrix(1, 6, h)
-
-    Group:
-        dsp-multirate-filtering
-    """
-    if not isinstance(up, int):
-        raise TypeError(f"Argument 'up' must be an integer, not {up}.")
-    if not up >= 1:
-        raise ValueError(f"Argument 'up' must be at least 1, not {up}.")
-    P = up
-
-    if not isinstance(down, int):
-        raise TypeError(f"Argument 'down' must be an integer, not {down}.")
-    if not down >= 1:
-        raise ValueError(f"Argument 'down' must be at least 1, not {down}.")
-    Q = down
-
-    taps = np.asarray(taps)
-
-    B = P if P > 1 else Q  # The number of polyphase branches
-    N = math.ceil(taps.size / B) * B  # Filter length
-
-    # Append zeros to the end of the taps so that the filter length is a multiple of B
-    taps_pad = np.pad(taps, (0, N - taps.size), mode="constant")
-
-    # Reshape the taps down the columns of H
-    H = taps_pad.reshape(-1, B).T
-
-    return H
 
 
 def _polyphase_interpolate(
@@ -375,7 +185,7 @@ class Interpolator(FIR):
             rate: The interpolation rate $r$.
             taps: The multirate filter design specification.
 
-                - `"kaiser"`: The multirate filter is designed using :func:`~sdr.multirate_taps()`
+                - `"kaiser"`: The multirate filter is designed using :func:`~sdr.design_multirate_fir()`
                   with arguments `rate` and 1.
                 - `"linear"`: The multirate filter is designed to linearly interpolate between samples.
                   The filter coefficients are a length-$2r$ linear ramp $\frac{1}{r} [0, ..., r-1, r, r-1, ..., 1]$.
@@ -402,16 +212,16 @@ class Interpolator(FIR):
             taps = np.asarray(taps)
         elif taps == "kaiser":
             self._method = "kaiser"
-            taps = multirate_taps(rate, 1)
+            taps = design_multirate_fir(rate, 1)
         elif taps == "linear":
             self._method = "linear"
-            taps = _multirate_taps_linear(rate)
+            taps = design_multirate_fir_linear(rate)
         elif taps == "linear-matlab":
             self._method = "linear-matlab"
-            taps = _multirate_taps_linear_matlab(rate)
+            taps = design_multirate_fir_linear_matlab(rate)
         elif taps == "zoh":
             self._method = "zoh"
-            taps = _multirate_taps_zoh(rate)
+            taps = design_multirate_fir_zoh(rate)
         else:
             raise ValueError(
                 f"Argument 'taps' must be 'kaiser', 'linear', 'linear-matlab', 'zoh', or an array-like, not {taps}."
@@ -763,7 +573,7 @@ class Decimator(FIR):
             rate: The decimation rate $r$.
             taps: The multirate filter design specification.
 
-                - `"kaiser"`: The multirate filter is designed using :func:`~sdr.multirate_taps()`
+                - `"kaiser"`: The multirate filter is designed using :func:`~sdr.design_multirate_fir()`
                   with arguments 1 and `rate`.
                 - `npt.ArrayLike`: The multirate filter feedforward coefficients $h_i$.
 
@@ -782,7 +592,7 @@ class Decimator(FIR):
             taps = np.asarray(taps)
         elif taps == "kaiser":
             self._method = "kaiser"
-            taps = multirate_taps(1, rate)
+            taps = design_multirate_fir(1, rate)
         else:
             raise ValueError(f"Argument 'taps' must be 'kaiser', or an array-like, not {taps}.")
 
@@ -1064,7 +874,7 @@ class Resampler(FIR):
             down: The decimation rate $Q$.
             taps: The multirate filter design specification.
 
-                - `"kaiser"`: The multirate filter is designed using :func:`~sdr.multirate_taps()`
+                - `"kaiser"`: The multirate filter is designed using :func:`~sdr.design_multirate_fir()`
                   with arguments `up` and `down`.
                 - `"linear"`: The multirate filter is designed to linearly interpolate between samples.
                   The filter coefficients are a length-$2P$ linear ramp $\frac{1}{P} [0, ..., P-1, P, P-1, ..., 1]$.
@@ -1099,22 +909,22 @@ class Resampler(FIR):
             taps = np.asarray(taps)
         elif taps == "kaiser":
             self._method = "kaiser"
-            taps = multirate_taps(up, down)
+            taps = design_multirate_fir(up, down)
         elif taps == "linear":
             if not up > 1:
                 raise ValueError(f"Argument 'up' must be greater than 1 to use the 'linear' method, not {up}.")
             self._method = "linear"
-            taps = _multirate_taps_linear(up)
+            taps = design_multirate_fir_linear(up)
         elif taps == "linear-matlab":
             if not up > 1:
                 raise ValueError(f"Argument 'up' must be greater than 1 to use the 'linear-matlab' method, not {up}.")
             self._method = "linear-matlab"
-            taps = _multirate_taps_linear_matlab(up)
+            taps = design_multirate_fir_linear_matlab(up)
         elif taps == "zoh":
             if not up > 1:
                 raise ValueError(f"Argument 'up' must be greater than 1 to use the 'zoh' method, not {up}.")
             self._method = "zoh"
-            taps = _multirate_taps_zoh(up)
+            taps = design_multirate_fir_zoh(up)
         else:
             raise ValueError(
                 f"Argument 'taps' must be 'kaiser', 'linear', 'linear-matlab', 'zoh', or an array-like, not {taps}."
